pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["nani1149 <nani1149@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
google-cloud-aiplatform = "^1.69.0"
apache-airflow = "^2.10.2"
apache-airflow-providers-google = "^10.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
